[refactor] presentation - Member관련 view, viewModel 분리
